Update issue 2921 Perimeter cells are now created natively as int64. arg-sorted events are also int 64.
Former-commit-id: 771ee26270eef2312ea4059b885b1cedc9fab4bd

+++ invest_natcap/scenic_quality/scenic_quality.py
@@ -447,9 +447,9 @@ def compute_viewshed(input_array, visibility_uri, in_structure_uri, \
             scenic_quality_cython_core.list_extreme_cell_angles(viewshed_shape, \
             v, max_dist)
         # arg_min, arg_center, arg_max
-        arg_min = np.argsort(add_events)
-        arg_max = np.argsort(remove_events)
-        arg_center = np.argsort(center_events)
+        arg_min = np.argsort(add_events).astype(np.int64)
+        arg_max = np.argsort(remove_events).astype(np.int64)
+        arg_center = np.argsort(center_events).astype(np.int64)
         # I and J are relative to the viewshed_shape. Make them absolute
         I += row_min
         J += col_min
